wip fusion for basic yaw. Sample time not yet working somehow.

+++ Core/Inc/timeManager.h
@@ -15,7 +15,7 @@ extern "C" {
 extern TIM_HandleTypeDef htim11;
 
 void timeInit(void);
-double getCurTime(void);
+float getCurTime(void);
 
 
 #ifdef __cplusplus
